Update to work with Debian

+++ setup.sh
@@ -1,28 +1,27 @@
 #!/usr/bin/env bash
 
 # Source bash config
-cat >> "$HOME/.bashrc" <<'EOF'
+cat >>"$HOME/.bashrc" <<'EOF'
 [[ -f "$HOME/dotfiles/bash/.bash_common" ]] && . "$HOME/dotfiles/bash/.bash_common"
 EOF
 
 mkdir -p "$HOME/Downloads/setup"
-version_num=$(grep VERSION_ID /etc/os-release | cut -d "=" -f 2 | sed -e s/\"//g)
-version_code=$(grep UBUNTU_CODENAME /etc/os-release | cut -d "=" -f 2)
+. /etc/os-release
 
 # Install required dependencies
-sudo apt install -y ca-certificates curl git gnupg software-properties-common
+sudo apt install -y ca-certificates curl git gnupg software-properties-common wget
 
 # Clone dotfiles and setup symlinks
 git clone https://github.com/kdien/dotfiles.git "$HOME/dotfiles"
 configs=(
-    alacritty
-    nvim
-    powershell
-    tmux
-    wezterm
+  alacritty
+  nvim
+  powershell
+  tmux
+  wezterm
 )
 for config in "${configs[@]}"; do
-    ln -sf "$HOME/dotfiles/$config" "$HOME/.config/$config"
+  ln -sf "$HOME/dotfiles/$config" "$HOME/.config/$config"
 done
 
 # Copy base git config
@@ -30,39 +29,35 @@ cp "$HOME/dotfiles/git/config" "$HOME/.gitconfig"
 
 # Configure GNOME settings
 if command -v gnome-shell &>/dev/null; then
-    sudo apt install -y gnome-tweaks gnome-shell-extensions gnome-shell-extension-appindicator
-    ./config-gnome.sh
-    mkdir -p "$HOME/bin"
-    for file in "$HOME"/dotfiles/gnome/*; do
-        ln -sf "$file" "$HOME/bin/$(basename "$file")"
-    done
+  sudo apt install -y gnome-tweaks gnome-shell-extensions gnome-shell-extension-appindicator
+  ./config-gnome.sh
+  mkdir -p "$HOME/bin"
+  for file in "$HOME"/dotfiles/gnome/*; do
+    ln -sf "$file" "$HOME/bin/$(basename "$file")"
+  done
 fi
 
 # Install fonts
 for font in "$HOME"/dotfiles/fonts/*.tar.gz; do
-    name=$(basename "$font" | cut -d '.' -f 1)
-    tar -xf "$font"
-    sudo chown root:root ./*.ttf
-    sudo mkdir -p "/usr/share/fonts/$name"
-    sudo mv ./*.ttf "/usr/share/fonts/$name"
+  name=$(basename "$font" | cut -d '.' -f 1)
+  dest="$HOME/.local/share/fonts/$name"
+  mkdir -p "$dest"
+  tar -xf "$font" --directory="$dest"
 done
 
-curl -sSL https://github.com/ryanoasis/nerd-fonts/releases/latest/download/NerdFontsSymbolsOnly.tar.xz -o nf-symbols.tar.xz
-tar -xf nf-symbols.tar.xz --wildcards '*.ttf'
-sudo chown root:root ./*.ttf
-sudo mkdir -p /usr/share/fonts/nf-symbols
-sudo mv ./*.ttf /usr/share/fonts/nf-symbols
-rm -f nf-symbols.tar.xz
-
 # Enable additional repos
-sudo add-apt-repository -y universe multiverse restricted
-sudo apt update
+if [[ "$ID" == ubuntu ]]; then
+  sudo add-apt-repository -y universe multiverse restricted
+  sudo apt update
+fi
 
 # Remove bloat
+# shellcheck disable=SC2046
 sudo apt remove -y $(cat ./pkg.remove)
-sudo snap remove firefox
+[[ "$ID" == ubuntu ]] && sudo snap remove firefox
 
 # Install packages from repo
+# shellcheck disable=SC2046
 sudo apt install -y $(cat ./pkg.add)
 
 # Build and install Neovim
@@ -78,6 +73,7 @@ cd "$OGPWD" || return
 
 # Install webi packages
 curl -sS https://webi.sh/webi | sh
+# shellcheck disable=SC2046
 webi $(cat ./webi.add)
 
 # Set up interception-tools
@@ -105,16 +101,18 @@ sudo mkdir -p /etc/interception/udevmon.d
 sudo install -o root -g root -m 644 caps2esc.yaml /etc/interception/udevmon.d/caps2esc.yaml
 sudo systemctl enable --now udevmon
 
-# Install Firefox from Mozilla
-curl -L "https://download.mozilla.org/?product=firefox-latest-ssl&os=linux64&lang=en-CA" -o "$HOME/Downloads/firefox.tar.bz2"
-tar -xf "$HOME/Downloads/firefox.tar.bz2"
-sudo rm -rf /opt/firefox
-sudo mv firefox /opt/firefox
-sudo mkdir -p /usr/local/bin
-sudo ln -s /opt/firefox/firefox /usr/local/bin/firefox
-sudo install -o root -g root -m 644 desktop-entries/firefox.desktop /usr/local/share/applications/firefox.desktop
-rm -f "$HOME/Downloads/firefox.tar.bz2"
-echo MOZ_ENABLE_WAYLAND=1 | sudo tee -a /etc/environment
+# Set up Firefox APT repo from Mozilla
+sudo install -d -m 0755 /etc/apt/keyrings
+wget -q https://packages.mozilla.org/apt/repo-signing-key.gpg -O- | sudo tee /etc/apt/keyrings/packages.mozilla.org.asc >/dev/null
+echo 'deb [signed-by=/etc/apt/keyrings/packages.mozilla.org.asc] https://packages.mozilla.org/apt mozilla main' | sudo tee -a /etc/apt/sources.list.d/mozilla.list >/dev/null
+
+sudo tee /etc/apt/preferences.d/mozilla <<EOF
+Package: *
+Pin: origin packages.mozilla.org
+Pin-Priority: 1000
+EOF
+sudo apt update
+sudo apt install -y firefox
 
 # Install Brave browser
 sudo curl -fsSL https://brave-browser-apt-release.s3.brave.com/brave-browser-archive-keyring.gpg -o /usr/share/keyrings/brave-browser-archive-keyring.gpg
@@ -142,14 +140,14 @@ EOF
 # Add Insync repo and install
 sudo apt-key adv --keyserver keyserver.ubuntu.com --recv-keys ACCAF35C
 sudo tee /etc/apt/sources.list.d/insync.list <<EOF
-deb http://apt.insync.io/ubuntu $version_code non-free contrib
+deb http://apt.insync.io/$ID "${VERSION_CODENAME/\"//}" non-free contrib
 EOF
 sudo apt update
 sudo apt install -y insync
 for filemgr in nautilus dolphin; do
-    if command -v "$filemgr" &>/dev/null; then
-        sudo apt install -y insync-"$filemgr"
-    fi
+  if command -v "$filemgr" &>/dev/null; then
+    sudo apt install -y insync-"$filemgr"
+  fi
 done
 
 # Install Zoom
@@ -157,7 +155,12 @@ curl -sSL https://zoom.us/client/latest/zoom_amd64.deb -o "$HOME/Downloads/setup
 sudo apt install -yf "$HOME/Downloads/setup/zoom.deb"
 
 # Install WezTerm
-curl -sSLo "$HOME/Downloads/setup/wezterm.deb" "$(curl -sSLH 'Accept: application/vnd.github+json' https://api.github.com/repos/wez/wezterm/releases/latest | jq -r ".assets[] | select(.browser_download_url | match(\"Ubuntu$version_num.deb$\")) | .browser_download_url")"
+curl -sSLo "$HOME/Downloads/setup/wezterm.deb" \
+  "$(
+    curl -sSLH 'Accept: application/vnd.github+json' https://api.github.com/repos/wez/wezterm/releases/latest \
+      | jq -r \
+        ".assets[] | select(.browser_download_url | match(\"${ID^}${VERSION_ID/\"//}.deb$\")) | .browser_download_url"
+  )"
 sudo apt install -yf "$HOME/Downloads/setup/wezterm.deb"
 
 # Clean up
